# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentOS API",
    "description": "AgentOS gateway API contract.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "PlatformCompanyUser": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "user_id",
          "display_name",
          "email",
          "role_bundle",
          "status",
          "last_sign_in_at",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "user_id": {
            "type": "string"
          },
          "display_name": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role_bundle": {
            "type": "string",
            "enum": [
              "COMPANY_ADMIN",
              "OPERATOR",
              "VIEWER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "INVITED",
              "ACTIVE",
              "DEACTIVATED"
            ]
          },
          "last_sign_in_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CompanyUsersResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformCompanyUser"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error_code",
          "message",
          "retryable",
          "correlation_id"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "correlation_id": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/company/users": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyUsersResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "responses": {
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformCompanyUser"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}